refactor: #1188 address PR review feedback
- clearUserCache: warn not error (swallow is intentional, error level is misleading)
- UserRelationRow.relation: String -> PermissionRelation (SDN maps enum names directly)
- Neo4jPermissionRepository.applyShareBatch: remove redundant try/catch (log+rethrow only)
- NoDuplicateUserIds: move namespace -> user pkg (validates UserMembershipRole)
