+++ agentos/agentos-service/src/main/kotlin/io/whozoss/agentos/permissions/UserRelationRow.kt
@@ -5,9 +5,11 @@ package io.whozoss.agentos.permissions
  * [PermissionNodeNeo4jRepository.findRelationsForUsers].
  *
  * Spring Data Neo4j maps the `RETURN u.id AS userId, type(r) AS relation` columns
- * onto this interface automatically.
+ * onto this interface automatically. The `relation` column holds the Neo4j
+ * relationship type string (e.g. `ADMIN`, `MEMBER`), which SDN maps directly to
+ * [PermissionRelation] because the enum constant names match.
  */
 interface UserRelationRow {
     val userId: String
-    val relation: String
+    val relation: PermissionRelation
 }
